requirements.txt
Django==4.0.4
django-cors-headers==3.12.0
django-debug-toolbar==3.4.0
django-filter==21.1
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dnspython==2.2.1
iso8601==1.0.2
motor==3.0.0
mysqlclient==2.1.0
phonenumbers==8.13.0
pipenv==2022.4.8
psycopg2==2.9.3
pytest==7.1.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.28.0

